Migrate all ITrap to IModifier

+++ RhythmDoctor.Archipelago/Modifiers/Archipelago/IArchipelagoModifier.cs
@@ -6,7 +6,7 @@ internal interface IArchipelagoModifier
   /// Gets how the trap's strength should scale
   /// based on how many item traps we have.
   /// </summary>
-  /// <param name="num">Number of items we have for this trap.</param>
-  /// <param name="consumed">Number of items that should be consumed for the strength given.</param>
-  double GetScale(int num, out int consumed);
+  /// <param name="num">Number of items we have for this trap. Must be at least 1.</param>
+  /// <param name="consumed">Number of items that should be consumed for the strength given. Must be at least 1.</param>
+  float GetScale(int num, out int consumed);
 }
